Conflicts:
	admin/changedata.php
	admin/inc/basic.php
	admin/inc/common.php
	admin/settings.php
	admin/template/css.php
	admin/template/js/jquery.getsimple.js
	admin/template/sidebar-files.php
	admin/upload-uploadify.php
	admin/upload.php
	temp.htaccess

+++ admin/setup.php
@@ -115,8 +115,11 @@ if(isset($_POST['submitted'])) {
 		}
 
 		# create root .htaccess file
-		 if ( !function_exists('apache_get_modules') or in_arrayi('mod_rewrite',apache_get_modules()) ) {
-				$init = GSROOTPATH.'.htaccess';
+		 if ( !function_exists('apache_get_modules') or in_arrayi('mod_rewrite',apache_get_modules())) {
+		 	$temp = GSROOTPATH .'temp.htaccess';
+		 	$init = GSROOTPATH.'.htaccess';
+			
+			if(file_exists($temp)) {				
 				$temp_data = file_get_contents(GSROOTPATH .'temp.htaccess');
 				$temp_data = str_replace('**REPLACE**',tsl($path_parts), $temp_data);
 				$fp = fopen($init, 'w');
@@ -124,16 +127,17 @@ if(isset($_POST['submitted'])) {
 				fclose($fp);
 				if (!file_exists($init)) {
 					$kill .= sprintf(i18n_r('ROOT_HTACCESS_ERROR'), 'temp.htaccess', '**REPLACE**', tsl($path_parts)) . '<br />';
-				} else {
-					unlink(GSROOTPATH .'temp.htaccess');
+				} else if(file_exists($temp)){
+					unlink($temp);
 				}
-		}
+			}	
+		} 
 	
 		# create gsconfig.php if it doesn't exist yet.
 		$init = GSROOTPATH.'gsconfig.php';
 		$temp = GSROOTPATH.'temp.gsconfig.php';
 		if (file_exists($init)) {
-			unlink($temp);
+			if(file_exists($temp)) unlink($temp);
 			if (file_exists($temp)) {
 				$kill .= sprintf(i18n_r('REMOVE_TEMPCONFIG_ERROR'), 'temp.gsconfig.php') . '<br />';
 			}
